Canonicalize after search; more reliably exclude.
Turns out it's hard to control the exact output of find; sometimes it
outputs ./foo/bar, sometimes it outputs foo/bar. By canonicalizing
find's output instead of its input, we fix this.

We also need to canonicalize the hardlink search exclusion exprs.
Because you may pass multiple search dirs, and the exclusion is relative
to each search dir, we canonicalize the target dirs relative to the
search dirs.